Update to ns-3-dev

+++ wscript
@@ -34,7 +34,7 @@ tests_enabled    = False
 
 # Bug 1868:  be conservative about -Wstrict-overflow for optimized builds
 # on older compilers; it can generate spurious warnings.  
-cc_version_warn_strict_overflow = ('4', '8', '2')
+gcc_version_warn_strict_overflow = ('4', '8', '2')
 
 # Bug 2181:  clang warnings about unused local typedefs and potentially
 # evaluated expressions affecting darwin clang/LLVM version 7.0.0 (Xcode 7)
@@ -146,6 +146,11 @@ def options(opt):
     opt.load('cflags')
     opt.load('gnu_dirs')
 
+    opt.add_option('--check-config',
+                   help=('Print the current configuration.'),
+                   action="store_true", default=False,
+                   dest="check_config")
+    
     opt.add_option('--cwd',
                    help=('Set the working directory for a program.'),
                    action="store", type="string", default=None,
@@ -232,12 +237,16 @@ def options(opt):
                          'but do not wait for ns-3 to finish the full build.'),
                    action="store_true", default=False,
                    dest='doxygen_no_build')
+    opt.add_option('--enable-des-metrics',
+                   help=('Log all events in a json file with the name of the executable (which must call CommandLine::Parse(argc, argv)'),
+                   action="store_true", default=False,
+                   dest='enable_desmetrics')
 
     # options provided in subdirectories
     opt.recurse('src')
     opt.recurse('bindings/python')
     opt.recurse('src/internet')
-
+    opt.recurse('contrib')
 
 def _check_compilation_flag(conf, flag, mode='cxx', linkflags=None):
     """
@@ -304,6 +313,29 @@ def _check_nonfatal(conf, *args, **kwargs):
     except conf.errors.ConfigurationError:
         return None
 
+# Write a summary of optional features status
+def print_config(env, phase='configure'):
+    if phase == 'configure':
+        profile = get_build_profile(env)
+    else:
+        profile = get_build_profile()
+        
+    print("---- Summary of optional NS-3 features:")
+    print("%-30s: %s%s%s" % ("Build profile", Logs.colors('GREEN'),
+                             profile, Logs.colors('NORMAL')))
+    bld = wutils.bld
+    print("%-30s: %s%s%s" % ("Build directory", Logs.colors('GREEN'),
+                             Options.options.out, Logs.colors('NORMAL')))
+    
+    
+    for (name, caption, was_enabled, reason_not_enabled) in sorted(env['NS3_OPTIONAL_FEATURES'], key=lambda s : s[1]):
+        if was_enabled:
+            status = 'enabled'
+            color = 'GREEN'
+        else:
+            status = 'not enabled (%s)' % reason_not_enabled
+            color = 'RED'
+        print("%-30s: %s%s%s" % (caption, Logs.colors(color), status, Logs.colors('NORMAL')))
 
 def configure(conf):
     conf.load('relocation', tooldir=['waf-tools'])
@@ -363,8 +395,8 @@ def configure(conf):
             if conf.check_compilation_flag('-march=native'):
                 env.append_value('CXXFLAGS', '-march=native') 
             env.append_value('CXXFLAGS', '-fstrict-overflow')
-            if conf.env['CC_VERSION'] == cc_version_warn_strict_overflow:
-                env.append_value('CXXFLAGS', '-Wstrict-overflow=5')
+            if conf.env['CC_VERSION'] >= gcc_version_warn_strict_overflow:
+                env.append_value('CXXFLAGS', '-Wstrict-overflow=2')
 
         if sys.platform == 'win32':
             env.append_value("LINKFLAGS", "-Wl,--enable-runtime-pseudo-reloc")
@@ -425,21 +457,46 @@ def configure(conf):
     conf.recurse('bindings/python')
 
     conf.recurse('src')
+    conf.recurse('contrib')
 
     # Set the list of enabled modules.
     if Options.options.enable_modules:
         # Use the modules explicitly enabled. 
-        conf.env['NS3_ENABLED_MODULES'] = ['ns3-'+mod for mod in
-                                           Options.options.enable_modules.split(',')]
+        _enabled_mods = []
+        _enabled_contrib_mods = []
+        for mod in Options.options.enable_modules.split(','): 
+            if mod in conf.env['NS3_MODULES'] and mod.startswith('ns3-'): 
+                _enabled_mods.append(mod)
+            elif 'ns3-' + mod in conf.env['NS3_MODULES']: 
+                _enabled_mods.append('ns3-' + mod)
+            elif mod in conf.env['NS3_CONTRIBUTED_MODULES'] and mod.startswith('ns3-'): 
+                _enabled_contrib_mods.append(mod)
+            elif 'ns3-' + mod in conf.env['NS3_CONTRIBUTED_MODULES']: 
+                _enabled_contrib_mods.append('ns3-' + mod)
+        conf.env['NS3_ENABLED_MODULES'] = _enabled_mods
+        conf.env['NS3_ENABLED_CONTRIBUTED_MODULES'] = _enabled_contrib_mods
+        
     else:
         # Use the enabled modules list from the ns3 configuration file.
         if modules_enabled[0] == 'all_modules':
             # Enable all modules if requested.
             conf.env['NS3_ENABLED_MODULES'] = conf.env['NS3_MODULES']
+            conf.env['NS3_ENABLED_CONTRIBUTED_MODULES'] = conf.env['NS3_CONTRIBUTED_MODULES']
         else:
             # Enable the modules from the list.
-            conf.env['NS3_ENABLED_MODULES'] = ['ns3-'+mod for mod in
-                                               modules_enabled]
+            _enabled_mods = []
+            _enabled_contrib_mods = []
+            for mod in modules_enabled: 
+                if mod in conf.env['NS3_MODULES'] and mod.startswith('ns3-'): 
+                    _enabled_mods.append(mod)
+                elif 'ns3-' + mod in conf.env['NS3_MODULES']: 
+                    _enabled_mods.append('ns3-' + mod)
+                elif mod in conf.env['NS3_CONTRIBUTED_MODULES'] and mod.startswith('ns3-'): 
+                    _enabled_contrib_mods.append(mod)
+                elif 'ns3-' + mod in conf.env['NS3_CONTRIBUTED_MODULES']: 
+                    _enabled_contrib_mods.append('ns3-' + mod)
+            conf.env['NS3_ENABLED_MODULES'] = _enabled_mods
+            conf.env['NS3_ENABLED_CONTRIBUTED_MODULES'] = _enabled_contrib_mods
 
     # Add the template module to the list of enabled modules that
     # should not be built if this is a static build on Darwin.  They
@@ -455,6 +512,8 @@ def configure(conf):
             conf.env['NS3_ENABLED_MODULES'].remove(not_built_name)
             if not conf.env['NS3_ENABLED_MODULES']:
                 raise WafError('Exiting because the ' + not_built + ' module can not be built and it was the only one enabled.')
+        elif not_built_name in conf.env['NS3_ENABLED_CONTRIBUTED_MODULES']:
+            conf.env['NS3_ENABLED_CONTRIBUTED_MODULES'].remove(not_built_name)
 
     conf.recurse('src/mpi')
 
@@ -556,6 +615,13 @@ def configure(conf):
     conf.report_optional_feature("libgcrypt", "Gcrypt library",
                                  conf.env.HAVE_GCRYPT, "libgcrypt not found: you can use libgcrypt-config to find its location.")
 
+    why_not_desmetrics = "defaults to disabled"
+    if Options.options.enable_desmetrics:
+        conf.env['ENABLE_DES_METRICS'] = True
+        env.append_value('DEFINES', 'ENABLE_DES_METRICS')
+        why_not_desmetrics = "option --enable-des-metrics selected"
+    conf.report_optional_feature("DES Metrics", "DES Metrics event collection", conf.env['ENABLE_DES_METRICS'], why_not_desmetrics)
+
 
     # for compiling C code, copy over the CXX* flags
     conf.env.append_value('CCFLAGS', conf.env['CXXFLAGS'])
@@ -571,8 +637,7 @@ def configure(conf):
     add_gcc_flag('-Wno-error=deprecated-declarations')
     add_gcc_flag('-fstrict-aliasing')
     add_gcc_flag('-Wstrict-aliasing')
-    add_gcc_flag('-Wno-error=maybe-uninitialized')
-    
+
     try:
         conf.find_program('doxygen', var='DOXYGEN')
     except WafError:
@@ -587,24 +652,8 @@ def configure(conf):
             value = shlex.split(os.environ[envvar])
             conf.env.append_value(confvar, value)
 
-    # Write a summary of optional features status
-    print("---- Summary of optional NS-3 features:")
-    print("%-30s: %s%s%s" % ("Build profile", Logs.colors('GREEN'),
-                             Options.options.build_profile, Logs.colors('NORMAL')))
-    bld = wutils.bld
-    print("%-30s: %s%s%s" % ("Build directory", Logs.colors('GREEN'),
-                             Options.options.out, Logs.colors('NORMAL')))
+    print_config(env)
     
-    
-    for (name, caption, was_enabled, reason_not_enabled) in sorted(conf.env['NS3_OPTIONAL_FEATURES'], key=lambda s : s[1]):
-        if was_enabled:
-            status = 'enabled'
-            color = 'GREEN'
-        else:
-            status = 'not enabled (%s)' % reason_not_enabled
-            color = 'RED'
-        print("%-30s: %s%s%s" % (caption, Logs.colors(color), status, Logs.colors('NORMAL')))
-
 
 class SuidBuild_task(Task.Task):
     """task that makes a binary Suid
@@ -697,7 +746,8 @@ def add_examples_programs(bld):
             return
 
 def add_scratch_programs(bld):
-    all_modules = [mod[len("ns3-"):] for mod in bld.env['NS3_ENABLED_MODULES']]
+    all_modules = [mod[len("ns3-"):] for mod in bld.env['NS3_ENABLED_MODULES'] + bld.env['NS3_ENABLED_CONTRIBUTED_MODULES']]
+
     try:
         for filename in os.listdir("scratch"):
             if filename.startswith('.') or filename == 'CVS':
@@ -748,12 +798,9 @@ def _find_ns3_module(self, name):
             return obj
     raise KeyError(name)
 
-
-def build(bld):
-    env = bld.env
-
-    if Options.options.check_profile:
-        # Parse the waf lockfile generated by latest 'configure' operation
+# Parse the waf lockfile generated by latest 'configure' operation
+def get_build_profile(env=None):
+    if env == None:
         lockfile = os.environ.get('WAFLOCK', '.lock-waf_%s_build' % sys.platform)
         profile = "not found"
         with open(lockfile, "r") as f:
@@ -765,7 +812,21 @@ def build(bld):
                         optkey,optval = x.split(':')
                         if (optkey.lstrip() == '\'build_profile\''):
                             profile = str(optval.lstrip()).replace("'","")
-        print("Build profile: %s" % profile)
+    else:
+        profile = Options.options.build_profile
+    return profile
+
+def build(bld):
+    env = bld.env
+
+    if Options.options.check_config:
+        print_config(env, 'build')
+    else:
+        if Options.options.check_profile:
+            profile = get_build_profile()
+            print("Build profile: %s" % profile)
+        
+    if Options.options.check_profile or Options.options.check_config:
         raise SystemExit(0)
         return
 
@@ -802,18 +863,21 @@ def build(bld):
 
     # process subfolders from here
     bld.recurse('src')
+    bld.recurse('contrib')
 
     # If modules have been enabled, then set lists of enabled modules
     # and enabled module test libraries.
-    if env['NS3_ENABLED_MODULES']:
+    if env['NS3_ENABLED_MODULES'] or env['NS3_ENABLED_CONTRIBUTED_MODULES']:
+
         modules = env['NS3_ENABLED_MODULES']
+        contribModules = env['NS3_ENABLED_CONTRIBUTED_MODULES']
 
         # Find out about additional modules that need to be enabled
         # due to dependency constraints.
         changed = True
         while changed:
             changed = False
-            for module in modules:
+            for module in modules + contribModules:
                 module_obj = bld.get_tgen_by_name(module)
                 if module_obj is None:
                     raise ValueError("module %s not found" % module)
@@ -821,31 +885,35 @@ def build(bld):
                 for dep in module_obj.use:
                     if not dep.startswith('ns3-'):
                         continue
-                    if dep not in modules:
-                        modules.append(dep)
+                    if dep not in modules and dep not in contribModules:
+                        if dep in env['NS3_MODULES']: modules.append(dep)
+                        elif dep in env['NS3_CONTRIBUTED_MODULES']: contribModules.append(dep)
                         changed = True
 
         env['NS3_ENABLED_MODULES'] = modules
+
+        env['NS3_ENABLED_CONTRIBUTED_MODULES'] = contribModules
 
         # If tests are being built, then set the list of the enabled
         # module test libraries.
         if env['ENABLE_TESTS']:
             for (mod, testlib) in bld.env['NS3_MODULES_WITH_TEST_LIBRARIES']:
-                if mod in bld.env['NS3_ENABLED_MODULES']:
+                if mod in bld.env['NS3_ENABLED_MODULES'] or mod in bld.env['NS3_ENABLED_CONTRIBUTED_MODULES']:
                     bld.env.append_value('NS3_ENABLED_MODULE_TEST_LIBRARIES', testlib)
 
     add_examples_programs(bld)
     add_scratch_programs(bld)
 
-    if env['NS3_ENABLED_MODULES']:
+    if env['NS3_ENABLED_MODULES'] or env['NS3_ENABLED_CONTRIBUTED_MODULES']:
         modules = env['NS3_ENABLED_MODULES']
+        contribModules = env['NS3_ENABLED_CONTRIBUTED_MODULES']
 
         # Exclude the programs other misc task gens that depend on disabled modules
         for obj in list(bld.all_task_gen):
 
             # check for ns3moduleheader_taskgen
             if 'ns3moduleheader' in getattr(obj, "features", []):
-                if ("ns3-%s" % obj.module) not in modules:
+                if ("ns3-%s" % obj.module) not in modules and ("ns3-%s" % obj.module) not in contribModules:
                     obj.mode = 'remove' # tell it to remove headers instead of installing
 
             # check for programs
@@ -853,7 +921,7 @@ def build(bld):
                 # this is an NS-3 program (bld.create_ns3_program)
                 program_built = True
                 for dep in obj.ns3_module_dependencies:
-                    if dep not in modules: # prog. depends on a module that isn't enabled?
+                    if dep not in modules and dep not in contribModules: # prog. depends on a module that isn't enabled?
                         bld.exclude_taskgen(obj)
                         program_built = False
                         break
@@ -874,38 +942,42 @@ def build(bld):
                     bld.env.append_value('NS3_RUNNABLE_PROGRAMS', object_relative_path)
 
             # disable the modules themselves
-            if hasattr(obj, "is_ns3_module") and obj.name not in modules:
+            if hasattr(obj, "is_ns3_module") and obj.name not in modules and obj.name not in contribModules:
                 bld.exclude_taskgen(obj) # kill the module
 
             # disable the module test libraries
             if hasattr(obj, "is_ns3_module_test_library"):
-                if not env['ENABLE_TESTS'] or (obj.module_name not in modules):
+                if not env['ENABLE_TESTS'] or ((obj.module_name not in modules) and (obj.module_name not in contribModules)) :
                     bld.exclude_taskgen(obj) # kill the module test library
 
             # disable the ns3header_taskgen
             if 'ns3header' in getattr(obj, "features", []):
-                if ("ns3-%s" % obj.module) not in modules:
+                if ("ns3-%s" % obj.module) not in modules and ("ns3-%s" % obj.module) not in contribModules:
                     obj.mode = 'remove' # tell it to remove headers instead of installing 
 
             # disable the ns3privateheader_taskgen
             if 'ns3privateheader' in getattr(obj, "features", []):
-                if ("ns3-%s" % obj.module) not in modules:
+                if ("ns3-%s" % obj.module) not in modules and ("ns3-%s" % obj.module) not in contribModules:
+
                     obj.mode = 'remove' # tell it to remove headers instead of installing 
 
             # disable pcfile taskgens for disabled modules
             if 'ns3pcfile' in getattr(obj, "features", []):
-                if obj.module not in bld.env.NS3_ENABLED_MODULES:
+                if obj.module not in bld.env.NS3_ENABLED_MODULES and obj.module not in bld.env.NS3_ENABLED_CONTRIBUTED_MODULES:
                     bld.exclude_taskgen(obj)
 
 
     if env['NS3_ENABLED_MODULES']:
         env['NS3_ENABLED_MODULES'] = list(modules)
 
+    if env['NS3_ENABLED_CONTRIBUTED_MODULES']:
+        env['NS3_ENABLED_CONTRIBUTED_MODULES'] = list(contribModules)
+
     # Determine which scripts will be runnable.
     for (script, dependencies) in bld.env['NS3_SCRIPT_DEPENDENCIES']:
         script_runnable = True
         for dep in dependencies:
-            if dep not in modules:
+            if dep not in modules and dep not in contribModules:
                 script_runnable = False
                 break
 
@@ -942,6 +1014,8 @@ def build(bld):
         for gen in bld.all_task_gen:
             if type(gen).__name__ in ['ns3header_taskgen', 'ns3privateheader_taskgen', 'ns3moduleheader_taskgen']:
                 gen.post()
+
+    if Options.options.run or Options.options.pyrun:
         bld.env['PRINT_BUILT_MODULES_AT_END'] = False 
 
     if Options.options.doxygen_no_build:
@@ -983,7 +1057,7 @@ def shutdown(ctx):
         print()
         print('Modules built:')
         names_without_prefix = []
-        for name in env['NS3_ENABLED_MODULES']:
+        for name in env['NS3_ENABLED_MODULES'] + env['NS3_ENABLED_CONTRIBUTED_MODULES']:
             name1 = name[len('ns3-'):]
             if name not in env.MODULAR_BINDINGS_MODULES:
                 name1 += " (no Python)"
